+++ tests/integration_tests/test_integration_simple_edge_system_from_json.py
@@ -60,6 +60,9 @@ class IntegrationTestSimpleEdgeSystemFromJson(IntegrationTestSimpleEdgeSystemBas
     def test_calculation_graph(self):
         self.run_test_calculation_graph()
 
+    def test_check_all_calculus_graph_dependencies_consistencies(self):
+        self.run_test_check_all_calculus_graph_dependencies_consistencies()
+
     # SYSTEM <=> JSON
 
     def test_json_to_system(self):