+++ tests/integration_tests/test_integration_simple_system_from_json.py
@@ -56,6 +56,9 @@ class IntegrationTestSimpleSystemFromJson(IntegrationTestSimpleSystemBaseClass):
     def test_object_relationship_graph(self):
         self.run_test_object_relationship_graph()
 
+    def test_check_all_calculus_graph_dependencies_consistencies(self):
+        self.run_test_check_all_calculus_graph_dependencies_consistencies()
+
     # SYSTEM <=> JSON
 
     def test_system_to_json(self):